clear draft from cache if discarded
