+++ app/src/views/features/sessions/SessionViewer/SaveRecordingConfigPopup.tsx
@@ -14,7 +14,6 @@ import {
   prepareSessionToExport,
 } from "./sessionEventsUtils";
 import { getSessionRecordingMetaData, getSessionRecordingEvents } from "store/features/session-recording/selectors";
-import { StorageService } from "init";
 import { toast } from "utils/Toast";
 import { getUserAuthDetails, getUserAttributes } from "store/selectors";
 import { getCurrentlyActiveWorkspace } from "store/features/teams/selectors";
@@ -32,12 +31,13 @@ import { submitAttrUtil } from "utils/AnalyticsUtils";
 interface Props {
   onClose: (e?: React.MouseEvent) => void;
   setIsSaveSessionClicked?: (value: boolean) => void;
+  clearSavedDraft?: () => void;
 }
 
 const { ACTION_LABELS: AUTH_ACTION_LABELS } = APP_CONSTANTS.AUTH;
 const defaultDebugInfo: CheckboxValueType[] = [DebugInfo.INCLUDE_NETWORK_LOGS, DebugInfo.INCLUDE_CONSOLE_LOGS];
 
-const SaveRecordingConfigPopup: React.FC<Props> = ({ onClose, setIsSaveSessionClicked }) => {
+const SaveRecordingConfigPopup: React.FC<Props> = ({ onClose, setIsSaveSessionClicked, clearSavedDraft }) => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const { tabId } = useParams();
@@ -82,14 +82,6 @@ const SaveRecordingConfigPopup: React.FC<Props> = ({ onClose, setIsSaveSessionCl
     [userAttributes?.num_sessions_saved_online, userAttributes?.num_sessions_saved_offline]
   );
 
-  const handleClearDraftSessionCache = useCallback(async () => {
-    const draftSessions = await StorageService().getRecord(APP_CONSTANTS.DRAFT_SESSIONS);
-    if (tabId in draftSessions) {
-      delete draftSessions[tabId];
-      StorageService().saveRecord({ [APP_CONSTANTS.DRAFT_SESSIONS]: { ...draftSessions } });
-    }
-  }, [tabId]);
-
   const saveDraftSession = useCallback(
     (e: React.MouseEvent) => {
       if (!user?.loggedIn) {
@@ -129,7 +121,7 @@ const SaveRecordingConfigPopup: React.FC<Props> = ({ onClose, setIsSaveSessionCl
       ).then((response) => {
         if (response?.success) {
           onClose();
-          handleClearDraftSessionCache();
+          clearSavedDraft?.();
           toast.success("Recording saved successfully");
           trackDraftSessionSaved(
             sessionRecordingMetadata?.sessionAttributes?.duration,
@@ -157,7 +149,7 @@ const SaveRecordingConfigPopup: React.FC<Props> = ({ onClose, setIsSaveSessionCl
       sessionEvents,
       sessionRecordingMetadata,
       setIsSaveSessionClicked,
-      handleClearDraftSessionCache,
+      clearSavedDraft,
       dispatch,
       navigate,
       onClose,
